+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.7.4] - 2023-06-10
+
+### Fixed
+
+- Improved `<Popper />` integration.
+- Fixed `aria-activedescendant` value.
+- Fixed double focus inside of popper.
+
 ## [3.7.3] - 2023-05-31
 
 ### Changed